Flappy Chicken - Problem set 0
Project on Scratch. My project is a game in which the chicken has to avoid touching the pipes that are coming towards it. for this the user uses the up and down arrows to move the chicken up and down, respectively. This is the link: https://scratch.mit.edu/projects/868104972